[FE] refactor: 기본 드롭다운에서 재사용이 가능하도록 훅 분리 (#29)
