--- src/frontend/src/hooks/empty.md
@@ -1,1 +1,0 @@
-폴더 생성을 위해 만든 빈 파일입니다.
